renamed entities folder to models folder, changed all imports from entities to models
