+++ angular/src/app/add-location/add-location.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../api.service';
-import { ProviderLocation } from 'src/Entities/location';
+import { ProviderLocation } from 'src/models/location';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
