+++ angular/src/app/app.module.ts
@@ -2,9 +2,9 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule, Pipe } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { Room } from 'src/Entities/room';
-import { Provider } from 'src/Entities/provider';
-import { ProviderLocation } from 'src/Entities/location';
+import { Room } from 'src/models/room';
+import { Provider } from 'src/models/provider';
+import { ProviderLocation } from 'src/models/location';
 import { NavComponent } from './nav/nav.component';
 import { HomeComponent } from './home/home.component';
 import { LoginComponent } from './login/login.component';
