+++ angular/src/models/room.spec.ts
