+++ angular/src/models/room.ts
